babe: move randomness utilities to its own file
